RetinaFace_Pytorch/requirements.txt
scikit_image==0.15.0
numpy==1.16.4
terminaltables==3.1.0
torch==1.1.0
tqdm==4.32.1
opencv_python==4.1.0.25
torchvision==0.3.0
Pillow==6.1.0
tensorboardX==1.8
